+++ ModelContextProtocol/McpServer.cs
@@ -15,27 +15,27 @@ namespace TiaMcpServer.ModelContextProtocol
 
         #region portal
 
-        [McpServerTool, Description("Connect to TIA Portal")]
+        [McpServerTool, Description("Connect to TIA-Portal")]
         public static string ConnectPortal()
         {
             try
             {
                 if (_portal.ConnectPortal())
                 {
-                    return JsonRpcMessageWrapper.ToJson(1, "Connected to TIA Portal successfully.");
+                    return JsonRpcMessageWrapper.ToJson(1, "Connected to TIA-Portal successfully.");
                 }
                 else
                 {
-                    return JsonRpcMessageWrapper.ToJson(1, false, "Failed to connect to TIA Portal.");
+                    return JsonRpcMessageWrapper.ToJson(1, false, "Failed to connect to TIA-Portal.");
                 }
             }
             catch (Exception ex)
             {
-                return JsonRpcMessageWrapper.ToJson(1, false, $"Error connecting to TIA Portal: {ex.Message}");
+                return JsonRpcMessageWrapper.ToJson(1, false, $"Error connecting to TIA-Portal: {ex.Message}");
             }
         }
 
-        [McpServerTool, Description("Check if TIA Portal is connected")]
+        [McpServerTool, Description("Check if TIA-Portal is connected")]
         public static string IsConnected()
         {
             try
@@ -49,17 +49,17 @@ namespace TiaMcpServer.ModelContextProtocol
             }
         }
 
-        [McpServerTool, Description("Disconnect from TIA Portal")]
+        [McpServerTool, Description("Disconnect from TIA-Portal")]
         public static string DisconnectPortal()
         {
             try
             {
                 _portal.DisconnectPortal();
-                return JsonRpcMessageWrapper.ToJson(1, "Disconnected from TIA Portal successfully.");
+                return JsonRpcMessageWrapper.ToJson(1, "Disconnected from TIA-Portal successfully.");
             }
             catch (Exception ex)
             {
-                return JsonRpcMessageWrapper.ToJson(1, false, $"Error disconnecting from TIA Portal: {ex.Message}");
+                return JsonRpcMessageWrapper.ToJson(1, false, $"Error disconnecting from TIA-Portal: {ex.Message}");
             }
         }
 
@@ -67,7 +67,7 @@ namespace TiaMcpServer.ModelContextProtocol
 
         #region project/session
 
-        [McpServerTool, Description("Get list of open projects/local sessions")]
+        [McpServerTool, Description("Get list of open local projects/sessions")]
         public static string GetOpenProjects()
         {
             try
@@ -83,7 +83,7 @@ namespace TiaMcpServer.ModelContextProtocol
             }
         }
 
-        [McpServerTool, Description("Open a TIA Portal project/local session")]
+        [McpServerTool, Description("Open a TIA-Portal local project/session")]
         public static string OpenProject(string projectPath)
         {
             try
@@ -101,25 +101,23 @@ namespace TiaMcpServer.ModelContextProtocol
                 }
 
                 bool success = false;
-                // switch case extension
-                switch (extension)
-                {
-                    case ".ap20":
-                        success = _portal.OpenProject(projectPath);
-                        break;
 
-                    case ".als20":
-                        success = _portal.OpenSession(projectPath);
-                        break;
+                if (extension.StartsWith(".ap"))
+                {
+                    success = _portal.OpenProject(projectPath);
+                }
+                if (extension.StartsWith(".als"))
+                {
+                    success = _portal.OpenSession(projectPath);
                 }
 
                 if (success)
                 {
-                    return JsonRpcMessageWrapper.ToJson(1, "Project opened successfully.");
+                    return JsonRpcMessageWrapper.ToJson(1, "Project opened successfully");
                 }
                 else
                 {
-                    return JsonRpcMessageWrapper.ToJson(1, false, "Failed to open project.");
+                    return JsonRpcMessageWrapper.ToJson(1, false, "Failed to open project");
                 }
             }
             catch (Exception ex)
@@ -128,7 +126,7 @@ namespace TiaMcpServer.ModelContextProtocol
             }
         }
 
-        [McpServerTool, Description("Save the current TIA Portal project/local session")]
+        [McpServerTool, Description("Save the current TIA-Portal local project/session")]
         public static string SaveProject()
         {
             try
@@ -137,60 +135,61 @@ namespace TiaMcpServer.ModelContextProtocol
                 {
                     if (_portal.SaveSession())
                     {
-                        return JsonRpcMessageWrapper.ToJson(1, "Local session saved successfully.");
+                        return JsonRpcMessageWrapper.ToJson(1, "Local session saved successfully");
                     }
                     else
                     {
-                        return JsonRpcMessageWrapper.ToJson(1, false, "Failed to save local session.");
+                        return JsonRpcMessageWrapper.ToJson(1, false, "Failed to save local session");
                     }
                 }
-                else 
+                else
                 {
                     if (_portal.SaveProject())
                     {
-                        return JsonRpcMessageWrapper.ToJson(1, "Project saved successfully.");
+                        return JsonRpcMessageWrapper.ToJson(1, "Local project saved successfully");
                     }
                     else
                     {
-                        return JsonRpcMessageWrapper.ToJson(1, false, "Failed to save project.");
+                        return JsonRpcMessageWrapper.ToJson(1, false, "Failed to save project");
                     }
                 }
             }
             catch (Exception ex)
             {
-                return JsonRpcMessageWrapper.ToJson(1, false, $"Error saving project/local session: {ex.Message}");
+                return JsonRpcMessageWrapper.ToJson(1, false, $"Error saving local project/session: {ex.Message}");
             }
         }
 
-        [McpServerTool, Description("Save the current TIA Portal project with a new name")]
-        public static string SaveAsProject(string newProjectPath)
+        [McpServerTool, Description("Save current TIA-Portal project/session with a new name")]
+        public static string SaveAsProject(
+            [Description("newProjectPath: defines the new path where to save the project")] string newProjectPath)
         {
             try
             {
                 if (_portal.IsLocalSession)
                 {
-                    return JsonRpcMessageWrapper.ToJson(1, false, $"Cannot save local session as {newProjectPath}.");
+                    return JsonRpcMessageWrapper.ToJson(1, false, $"Cannot save local session as '{newProjectPath}'");
                 }
                 else
                 {
                     if (_portal.SaveAsProject(newProjectPath))
                     {
-                        return JsonRpcMessageWrapper.ToJson(1, $"Project saved successfully as {newProjectPath}.");
+                        return JsonRpcMessageWrapper.ToJson(1, $"Local project saved successfully as '{newProjectPath}'");
                     }
                     else
                     {
-                        return JsonRpcMessageWrapper.ToJson(1, false, $"Failed to save project as {newProjectPath}.");
+                        return JsonRpcMessageWrapper.ToJson(1, false, $"Failed to save local project as '{newProjectPath}'");
                     }
                 }
-                
+
             }
             catch (Exception ex)
             {
-                return JsonRpcMessageWrapper.ToJson(1, false, $"Error saving project/local session as {newProjectPath}: {ex.Message}");
+                return JsonRpcMessageWrapper.ToJson(1, false, $"Error saving local project/session as '{newProjectPath}': {ex.Message}");
             }
         }
 
-        [McpServerTool, Description("Close the current TIA Portal project")]
+        [McpServerTool, Description("Close the current TIA-Portal project/session")]
         public static string CloseProject()
         {
             try
@@ -199,19 +198,19 @@ namespace TiaMcpServer.ModelContextProtocol
                 {
                     _portal.CloseSession();
 
-                    return JsonRpcMessageWrapper.ToJson(1, "Local session closed successfully.");
+                    return JsonRpcMessageWrapper.ToJson(1, "Local session closed successfully");
                 }
                 else
                 {
                     _portal.CloseProject();
 
-                    return JsonRpcMessageWrapper.ToJson(1, "Project closed successfully.");
+                    return JsonRpcMessageWrapper.ToJson(1, "Local project closed successfully");
                 }
-                
+
             }
             catch (Exception ex)
             {
-                return JsonRpcMessageWrapper.ToJson(1, false, $"Error closing project/local session: {ex.Message}");
+                return JsonRpcMessageWrapper.ToJson(1, false, $"Error closing local project/session: {ex.Message}");
             }
         }
 
@@ -219,34 +218,24 @@ namespace TiaMcpServer.ModelContextProtocol
 
         #region devices
 
-        [McpServerTool, Description("Get the structure of the current project/local session")]
+        [McpServerTool, Description("Get the structure of current local project/session")]
         public static string GetStructure()
         {
             try
             {
-                return _portal.GetStructure();
+                var structure = _portal.GetStructure();
+
+                return JsonRpcMessageWrapper.ToJson(1, false, structure);
             }
             catch (Exception ex)
             {
-                return $"Error retrieving structure of the project/local session: {ex.Message}";
+                return JsonRpcMessageWrapper.ToJson(1, false, $"Error retrieving structure of current local the project/session: {ex.Message}");
             }
         }
 
-        [McpServerTool, Description("Get a list of devices in the project")]
-        public static List<string> GetDevices()
-        {
-            try
-            {
-                return _portal.GetDevices();
-            }
-            catch (Exception ex)
-            {
-                return [$"Error retrieving devices: {ex.Message}"];
-            }
-        }
-
-        [McpServerTool, Description("Get a device by its path in the project")]
-        public static string GetDevice(string devicePath)
+        [McpServerTool, Description("Get info from a device from the current project/session")]
+        public static string GetDeviceInfo(
+            [Description("devicePath: defines the path in the project structure to the device")] string devicePath)
         {
             try
             {
@@ -254,14 +243,55 @@ namespace TiaMcpServer.ModelContextProtocol
 
                 if (device == null)
                 {
-                    return $"Device '{devicePath}' not found.";
+                    return JsonRpcMessageWrapper.ToJson(1, false, $"Device '{devicePath}' not found.");
                 }
 
-                return device.Name;
+                return JsonRpcMessageWrapper.ToJson(1, false, device.Name);
             }
             catch (Exception ex)
             {
-                return $"Error retrieving device: {ex.Message}";
+                return JsonRpcMessageWrapper.ToJson(1, false, $"Error retrieving device '{devicePath}': {ex.Message}");
+            }
+        }
+
+        [McpServerTool, Description("Get info from a device item from the current project/session")]
+        public static string GetDeviceItemInfo(
+            [Description("deviceItemPath: defines the path in the project structure to the device item")] string deviceItemPath)
+        {
+            try
+            {
+                var deviceItem = _portal.GetDeviceItem(deviceItemPath);
+
+                if (deviceItem == null)
+                {
+                    return JsonRpcMessageWrapper.ToJson(1, false, $"Device item '{deviceItemPath}' not found.");
+                }
+
+                return JsonRpcMessageWrapper.ToJson(1, false, deviceItem.Name);
+            }
+            catch (Exception ex)
+            {
+                return JsonRpcMessageWrapper.ToJson(1, false, $"Error retrieving device item '{deviceItemPath}': {ex.Message}");
+            }
+        }
+
+
+        [McpServerTool, Description("Get a list of all devices in the project/session")]
+        public static string GetDevices()
+        {
+            try
+            {
+                var list = _portal.GetDevices();
+                if (list.Count == 0)
+                {
+                    return JsonRpcMessageWrapper.ToJson(1, "No devices found in the current project/session");
+                }
+
+                return JsonRpcMessageWrapper.ToJson(1, list);
+            }
+            catch (Exception ex)
+            {
+                return JsonRpcMessageWrapper.ToJson(1, false, $"Error retrieving devices: {ex.Message}");
             }
         }
 
@@ -269,8 +299,9 @@ namespace TiaMcpServer.ModelContextProtocol
 
         #region plc software
 
-        [McpServerTool, Description("Compile the software, which is given by softwarePath")]
-        public static string CompileSoftware(string softwarePath)
+        [McpServerTool, Description("Compile the plc software.")]
+        public static string CompileSoftware(
+            [Description("softwarePath: defines the path in the project structure to the plc software")] string softwarePath)
         {
             try
             {
@@ -278,156 +309,130 @@ namespace TiaMcpServer.ModelContextProtocol
 
                 if (!result.Equals("Error"))
                 {
-                    return JsonRpcMessageWrapper.ToJson(1, $"Software compiled with {result}.");
+                    return JsonRpcMessageWrapper.ToJson(1, $"Software '{softwarePath}' compiled with {result}");
                 }
                 else
                 {
-                    return JsonRpcMessageWrapper.ToJson(1, false, $"Failed to compile software with {result}");
+                    return JsonRpcMessageWrapper.ToJson(1, false, $"Failed to compile software '{softwarePath}' with {result}");
                 }
             }
             catch (Exception ex)
             {
-                return JsonRpcMessageWrapper.ToJson(1, false, $"Error compiling software: {ex.Message}");
+                return JsonRpcMessageWrapper.ToJson(1, false, $"Error compiling software '{softwarePath}': {ex.Message}");
             }
         }
 
         #endregion
 
-        #region code blocks
+        #region blocks
 
-        [McpServerTool, Description("Get a code block (FB, FC, OB) from the software, which is given by softwarePath/groupPath/blockName.")]
-        public static string GetCodeBlock(string softwarePath, string groupPath, string blockName)
+        [McpServerTool, Description("Get a block info, which is located in the plc software")]
+        public static string GetBlockInfo(
+            [Description("softwarePath: defines the path in the project structure to the plc software")] string softwarePath,
+            [Description("blockPath: defines the path in the project structure to the block")] string blockPath)
         {
             try
             {
-                return _portal.GetCodeBlock(softwarePath, groupPath, blockName);
-            }
-            catch (Exception ex)
-            {
-                return $"Error retrieving data block: {ex.Message}";
-            }
-        }
-
-        [McpServerTool, Description("Get a list of all code blocks (FB, FC, OB) from the software, which is given by softwarePath.")]
-        public static List<string> GetCodeBlocks(string softwarePath)
-        {
-            try
-            {
-                return _portal.GetCodeBlocks(softwarePath);
-            }
-            catch (Exception ex)
-            {
-                return [$"Error retrieving code blocks: {ex.Message}"];
-            }
-        }
-
-        [McpServerTool, Description("Export a code block (FB, FC, OB) given by softwarePath/groupPath/blockName to a specified exportPath.")]
-        public static string ExportCodeBlock(string softwarePath, string groupPath, string blockName, string exportPath)
-        {
-            try
-            {
-                if (_portal.ExportCodeBlock(softwarePath, groupPath, blockName, exportPath))
+                var block = _portal.GetBlock(softwarePath, blockPath);
+                if (block == null)
                 {
-                    return JsonRpcMessageWrapper.ToJson(1, "Code block exported successfully.");
+                    return JsonRpcMessageWrapper.ToJson(1, false, $"Block '{blockPath}' not found in software '{softwarePath}'.");
+                }
+                return JsonRpcMessageWrapper.ToJson(1, false, block.Name + ", " + block.Namespace + ", " + block.ProgrammingLanguage);
+            }
+            catch (Exception ex)
+            {
+                return JsonRpcMessageWrapper.ToJson(1, false, $"Error retrieving block '{blockPath}': {ex.Message}");
+            }
+        }
+
+        [McpServerTool, Description("Get a list of blocks, which are located in plc software")]
+        public static string GetBlocks(
+            [Description("softwarePath: defines the path in the project structure to the plc software")] string softwarePath,
+            [Description("regexName: defines the name or regular expression to find the block. Use empty string (default) to find all")] string regexName = "")
+        {
+            try
+            {
+                var list = _portal.GetBlocks(softwarePath, regexName);
+                if (list.Count == 0)
+                {
+                    return JsonRpcMessageWrapper.ToJson(1, $"No blocks found in software '{softwarePath}' with regex '{regexName}'.");
+                }
+
+                return JsonRpcMessageWrapper.ToJson(1, list);
+            }
+            catch (Exception ex)
+            {
+                return JsonRpcMessageWrapper.ToJson(1, false, $"Error retrieving code blocks: {ex.Message}");
+            }
+        }
+
+        [McpServerTool, Description("Export a block from plc software to file")]
+        public static string ExportBlock(
+            [Description("softwarePath: defines the path in the project structure to the plc software")] string softwarePath,
+            [Description("blockPath: defines the path in the project structure to the block")] string blockPath,
+            [Description("exportPath: defines the path where to export the block")] string exportPath)
+        {
+            try
+            {
+                if (_portal.ExportBlock(softwarePath, blockPath, exportPath))
+                {
+                    return JsonRpcMessageWrapper.ToJson(1, $"Block exported successfully from '{blockPath}' to '{exportPath}'");
                 }
                 else
                 {
-                    return JsonRpcMessageWrapper.ToJson(1, false, "Failed to export code block.");
+                    return JsonRpcMessageWrapper.ToJson(1, false, $"Failed to export block from '{blockPath}' to '{exportPath}'");
                 }
             }
             catch (Exception ex)
             {
-                return JsonRpcMessageWrapper.ToJson(1, false, $"Error exporting code block: {ex.Message}");
+                return JsonRpcMessageWrapper.ToJson(1, false, $"Error exporting block from '{blockPath}' to '{exportPath}': {ex.Message}");
             }
         }
 
-        [McpServerTool, Description("Export all code blocks (FB, FC, OB) from the software, given by softwarePath, to a specified path.")]
-        public static string ExportCodeBlocks(string softwarePath, string exportPath)
+        [McpServerTool, Description("Import a block file to plc software")]
+        public static string ImportBlock(
+            [Description("softwarePath: defines the path in the project structure to the plc software")] string softwarePath,
+            [Description("groupPath: defines the path in the project structure to the group, where to import the block")] string groupPath,
+            [Description("importPath: defines the path of the xml file from where to import the block")] string importPath)
         {
             try
             {
-                if (_portal.ExportCodeBlocks(softwarePath, exportPath))
+                if (_portal.ImportBlock(softwarePath, groupPath, importPath))
                 {
-                    return JsonRpcMessageWrapper.ToJson(1, "Code blocks exported successfully.");
+                    return JsonRpcMessageWrapper.ToJson(1, $"Block imported successfully from '{importPath}' to '{groupPath}'");
                 }
                 else
                 {
-                    return JsonRpcMessageWrapper.ToJson(1, false, "Failed to export code blocks.");
+                    return JsonRpcMessageWrapper.ToJson(1, false, $"Failed to import block from '{importPath}' to '{groupPath}'");
                 }
             }
             catch (Exception ex)
             {
-                return JsonRpcMessageWrapper.ToJson(1, false, $"Error exporting code blocks: {ex.Message}");
+                return JsonRpcMessageWrapper.ToJson(1, false, $"Error importing block from '{importPath}' to '{groupPath}': {ex.Message}");
             }
         }
 
-        #endregion
-
-        #region data blocks
-
-        [McpServerTool, Description("Get a data block (InstanceDB, GlobalDB, ArrayDB), given by softwarePath/groupPath/blockName.")]
-        public static string GetDataBlock(string softwarePath, string groupPath, string blockName)
+        [McpServerTool, Description("Export all blocks from the plc software to path")]
+        public static string ExportBlocks(
+            [Description("softwarePath: defines the path in the project structure to the plc software")] string softwarePath,
+            [Description("exportPath: defines the path where to export the blocks")] string exportPath,
+            [Description("regexName: defines the name or regular expression to find the block. Use empty string (default) to find all")] string regexName = "")
         {
             try
             {
-                return _portal.GetDataBlock(softwarePath, groupPath, blockName);
-            }
-            catch (Exception ex)
-            {
-                return $"Error retrieving data block: {ex.Message}";
-            }
-        }
-
-        [McpServerTool, Description("Get list of data blocks (InstanceDB, GlobalDB, ArrayDB) from software, given by softwarePath, in the project.")]
-        public static List<string> GetDataBlocks(string softwarePath)
-        {
-            try
-            {
-                return _portal.GetDataBlocks(softwarePath);
-            }
-            catch (Exception ex)
-            {
-                return [$"Error retrieving data blocks: {ex.Message}"];
-            }
-        }
-
-        [McpServerTool, Description("Export a data block given by softwarePath/groupPath/blockName to a specified exportPath (InstanceDB, GlobalDB, ArrayDB).")]
-        public static string ExportDataBlock(string softwarePath, string groupPath, string blockName, string exportPath)
-        {
-            try
-            {
-                if (_portal.ExportDataBlock(softwarePath, groupPath, blockName, exportPath))
+                if (_portal.ExportBlocks(softwarePath, exportPath, regexName))
                 {
-                    return JsonRpcMessageWrapper.ToJson(1, "Data block exported successfully.");
+                    return JsonRpcMessageWrapper.ToJson(1, "Blocks exported successfully");
                 }
                 else
                 {
-                    return JsonRpcMessageWrapper.ToJson(1, false, "Failed to export data block.");
+                    return JsonRpcMessageWrapper.ToJson(1, false, "Failed to export blocks");
                 }
             }
             catch (Exception ex)
             {
-                return JsonRpcMessageWrapper.ToJson(1, false, $"Error exporting data block: {ex.Message}");
-            }
-        }
-
-        [McpServerTool, Description("Export all data blocks (InstanceDB, GlobalDB, ArrayDB) from the software, given by softwarePath, to a specified path.")]
-        public static string ExportDataBlocks(string softwarePath, string exportPath)
-        {
-            try
-            {
-                if (_portal.ExportDataBlocks(softwarePath, exportPath))
-                {
-                    return JsonRpcMessageWrapper.ToJson(1, "Data blocks exported successfully.");
-                }
-                else
-                {
-                    return JsonRpcMessageWrapper.ToJson(1, false, "Failed to export data blocks.");
-                }
-            }
-            catch (Exception ex)
-            {
-                return JsonRpcMessageWrapper.ToJson(1, false, $"Error exporting data blocks: {ex.Message}");
+                return JsonRpcMessageWrapper.ToJson(1, false, $"Error exporting blocks: {ex.Message}");
             }
         }
 
@@ -435,25 +440,35 @@ namespace TiaMcpServer.ModelContextProtocol
 
         #region types
 
-        [McpServerTool, Description("Get a user defined type (UDT) from the software, which is given by softwarePath.")]
-        public static string GetUserDefinedType(string softwarePath, string typeName)
+        [McpServerTool, Description("Get a type info from the plc software")]
+        public static string GetTypeInfo(
+            [Description("softwarePath: defines the path in the project structure to the plc software")] string softwarePath,
+            [Description("typePath: defines the path in the project structure to the type")] string typePath)
         {
             try
             {
-                return _portal.GetUserDefinedType(softwarePath, typeName);
+                var type = _portal.GetType(softwarePath, typePath);
+                if (type == null)
+                {
+                    return JsonRpcMessageWrapper.ToJson(1, $"Type '{typePath}' not found in software '{softwarePath}'");
+                }
+
+                return JsonRpcMessageWrapper.ToJson(1, type.Name + ", " + type.Namespace);
             }
             catch (Exception ex)
             {
-                return $"Error retrieving user defined type: {ex.Message}";
+                return JsonRpcMessageWrapper.ToJson(1, $"Error retrieving user defined type '{typePath}': {ex.Message}");
             }
         }
 
-        [McpServerTool, Description("Get a list of all user defined types (UDT) from the software, which is given by softwarePath.")]
-        public static List<string> GetUserDefinedTypes(string softwarePath)
+        [McpServerTool, Description("Get a list of types from the plc software")]
+        public static List<string> GetTypes(
+            [Description("softwarePath: defines the path in the project structure to the plc software")] string softwarePath,
+            [Description("regexName: defines the name or regular expression to find the block. Use empty string (default) to find all")] string regexName = "")
         {
             try
             {
-                return _portal.GetUserDefinedTypes(softwarePath);
+                return _portal.GetTypes(softwarePath, regexName);
             }
             catch (Exception ex)
             {
@@ -461,43 +476,99 @@ namespace TiaMcpServer.ModelContextProtocol
             }
         }
 
-        [McpServerTool, Description("Export a user defined types (UDT) by name from the software, which is giveb by softwarePath.")]
-        public static string ExportUserDefinedType(string softwarePath, string exportPath, string typeName)
+        [McpServerTool, Description("Export a type from the plc software")]
+        public static string ExportType(
+            [Description("softwarePath: defines the path in the project structure to the plc software")] string softwarePath,
+            [Description("exportPath: defines the path where export the type")] string exportPath,
+            [Description("typePath: defines the path in the project structure to the type")] string typePath)
         {
             try
             {
-                if (_portal.ExportUserDefinedType(softwarePath, exportPath, typeName))
+                if (_portal.ExportType(softwarePath, typePath, exportPath))
                 {
-                    return JsonRpcMessageWrapper.ToJson(1, "User defined type exported successfully.");
+                    return JsonRpcMessageWrapper.ToJson(1, $"Type exported successfully from '{typePath}' to '{exportPath}'");
                 }
                 else
                 {
-                    return JsonRpcMessageWrapper.ToJson(1, false, "Failed to export user defined type.");
+                    return JsonRpcMessageWrapper.ToJson(1, false, $"Failed to export type from '{typePath}' to '{exportPath}'");
                 }
             }
             catch (Exception ex)
             {
-                return JsonRpcMessageWrapper.ToJson(1, false, $"Error exporting user defined type: {ex.Message}");
+                return JsonRpcMessageWrapper.ToJson(1, false, $"Error exporting type from '{typePath}' to '{exportPath}': {ex.Message}");
             }
         }
 
-        [McpServerTool, Description("Export all user defined types (UDT) from the software, which is giveb by softwarePath.")]
-        public static string ExportUserDefinedTypes(string softwarePath, string exportPath)
+        [McpServerTool, Description("Import a type from file into the plc software")]
+        public static string ImportType(
+            [Description("softwarePath: defines the path in the project structure to the plc software")] string softwarePath,
+            [Description("groupPath: defines the path in the project structure to the group, where to import the type")] string groupPath,
+            [Description("importPath: defines the path of the xml file from where to import the type")] string importPath)
         {
             try
             {
-                if (_portal.ExportUserDefinedTypes(softwarePath, exportPath))
+                if (_portal.ImportType(softwarePath, groupPath, importPath))
                 {
-                    return JsonRpcMessageWrapper.ToJson(1, "User defined types exported successfully.");
+                    return JsonRpcMessageWrapper.ToJson(1, $"Type imported successfully from '{importPath}' to '{groupPath}'");
                 }
                 else
                 {
-                    return JsonRpcMessageWrapper.ToJson(1, false, "Failed to export user defined types.");
+                    return JsonRpcMessageWrapper.ToJson(1, false, $"Failed to import type from '{importPath}' to '{groupPath}'");
                 }
             }
             catch (Exception ex)
             {
-                return JsonRpcMessageWrapper.ToJson(1, false, $"Error exporting user defined types: {ex.Message}");
+                return JsonRpcMessageWrapper.ToJson(1, false, $"Error importing type from '{importPath}' to '{groupPath}': {ex.Message}");
+            }
+        }
+
+        [McpServerTool, Description("Export types from the plc software to path")]
+        public static string ExportTypes(
+            [Description("softwarePath: defines the path in the project structure to the plc software")] string softwarePath,
+            [Description("exportPath: defines the path where to export the types")] string exportPath,
+            [Description("regexName: defines the name or regular expression to find the block. Use empty string (default) to find all")] string regexName = "")
+        {
+            try
+            {
+                if (_portal.ExportTypes(softwarePath, exportPath, regexName))
+                {
+                    return JsonRpcMessageWrapper.ToJson(1, "Types exported successfully");
+                }
+                else
+                {
+                    return JsonRpcMessageWrapper.ToJson(1, false, "Failed to export types");
+                }
+            }
+            catch (Exception ex)
+            {
+                return JsonRpcMessageWrapper.ToJson(1, false, $"Error exporting types: {ex.Message}");
+            }
+        }
+
+        #endregion
+
+        #region documents
+
+        [McpServerTool, Description("Export as documents (.s7dcl/.s7res) from a block in the plc software to path")]
+        public static string ExportAsDocuments(
+            [Description("softwarePath: defines the path in the project structure to the plc software")] string softwarePath,
+            [Description("blockPath: defines the path in the project structure to the block")] string blockPath,
+            [Description("exportPath: defines the path where to export the documents")] string exportPath)
+        {
+            try
+            {
+                if (_portal.ExportAsDocuments(softwarePath, blockPath, exportPath))
+                {
+                    return JsonRpcMessageWrapper.ToJson(1, $"Documents exported successfully from '{blockPath}' to '{exportPath}'");
+                }
+                else
+                {
+                    return JsonRpcMessageWrapper.ToJson(1, false, $"Failed to export documents from '{blockPath}' to '{exportPath}'");
+                }
+            }
+            catch (Exception ex)
+            {
+                return JsonRpcMessageWrapper.ToJson(1, false, $"Error exporting documents from '{blockPath}' to '{exportPath}': {ex.Message}");
             }
         }